final smoll fix for today

+++ project/helpers.py
@@ -18,7 +18,7 @@ right_half_center = mn.ORIGIN + mn.RIGHT * mn.config.frame_x_radius / 2
 left_half_center = mn.ORIGIN + mn.LEFT * mn.config.frame_x_radius / 2
 
 
-def fade_out(scene: mn.Scene, run_time: int = 1):
+def fade_out(scene: mn.Scene, run_time: float = 1):
     anims = []
     for mobj in scene.mobjects:
         anims.append(mn.FadeOut(mobj, run_time=run_time))
